+++ x/dex/client/testutil/cli_helpers.go
@@ -1,4 +1,4 @@
-package testutil
+package cli_test
 
 import (
 	"fmt"